# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arete – Solana Pay invoicing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NonceResponse {
  nonce: string;
}

export interface WalletVerifyInput {
  wallet: string;
  signature: string;
  nonce: string;
}

export interface AuthUser {
  wallet: string;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface Invoice {
  id: string;
  clientName: string;
  amountUsdc: number;
  recipientWallet: string;
  referenceId: string;
  ownerWallet: string;
  status: InvoiceStatus;
  solanaPayUrl: string;
  /** @nullable */
  txSignature?: string | null;
  createdAt: string;
  /** @nullable */
  paidAt?: string | null;
}

export interface InvoiceInput {
  /** @minLength 1 */
  clientName: string;
  /** @minimum 0.01 */
  amountUsdc: number;
  /** @minLength 32 */
  recipientWallet: string;
}

export interface InvoiceStats {
  totalInvoices: number;
  paidInvoices: number;
  pendingInvoices: number;
  totalPaidUsdc: number;
  totalPendingUsdc: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetNonceParams = {
wallet: string;
};