# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arete – Solana Pay invoicing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Sign In With Solana authentication
  - name: invoices
    description: Invoice management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/nonce:
    get:
      operationId: getNonce
      tags: [auth]
      summary: Get a one-time nonce for wallet signing
      parameters:
        - name: wallet
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Nonce to sign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonceResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify:
    post:
      operationId: verifyWallet
      tags: [auth]
      summary: Verify wallet signature and issue session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletVerifyInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated wallet
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Clear session
      responses:
        "204":
          description: Logged out

  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices for the authenticated wallet
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice and generate a Solana Pay link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice with Solana Pay URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /invoices/stats:
    get:
      operationId: getInvoiceStats
      tags: [invoices]
      summary: Dashboard stats for the authenticated wallet
      responses:
        "200":
          description: Aggregated invoice statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceStats"

  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get a single invoice by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /invoices/{id}/watch:
    post:
      operationId: watchInvoice
      tags: [invoices]
      summary: Trigger on-chain payment detection for this invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated invoice (may now be paid)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NonceResponse:
      type: object
      required:
        - nonce
      properties:
        nonce:
          type: string

    WalletVerifyInput:
      type: object
      required:
        - wallet
        - signature
        - nonce
      properties:
        wallet:
          type: string
        signature:
          type: string
        nonce:
          type: string

    AuthUser:
      type: object
      required:
        - wallet
      properties:
        wallet:
          type: string

    Invoice:
      type: object
      required:
        - id
        - clientName
        - amountUsdc
        - recipientWallet
        - referenceId
        - status
        - solanaPayUrl
        - createdAt
        - ownerWallet
      properties:
        id:
          type: string
        clientName:
          type: string
        amountUsdc:
          type: number
        recipientWallet:
          type: string
        referenceId:
          type: string
        ownerWallet:
          type: string
        status:
          type: string
          enum: [pending, paid]
        solanaPayUrl:
          type: string
        txSignature:
          type: ["string", "null"]
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]

    InvoiceInput:
      type: object
      required:
        - clientName
        - amountUsdc
        - recipientWallet
      properties:
        clientName:
          type: string
          minLength: 1
        amountUsdc:
          type: number
          minimum: 0.01
        recipientWallet:
          type: string
          minLength: 32

    InvoiceStats:
      type: object
      required:
        - totalInvoices
        - paidInvoices
        - pendingInvoices
        - totalPaidUsdc
        - totalPendingUsdc
      properties:
        totalInvoices:
          type: integer
        paidInvoices:
          type: integer
        pendingInvoices:
          type: integer
        totalPaidUsdc:
          type: number
        totalPendingUsdc:
          type: number

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string