# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arete – Solana Pay invoicing API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceStatus } from './invoiceStatus';

export interface Invoice {
  id: string;
  clientName: string;
  amountUsdc: number;
  recipientWallet: string;
  referenceId: string;
  ownerWallet: string;
  status: InvoiceStatus;
  solanaPayUrl: string;
  /** @nullable */
  txSignature?: string | null;
  createdAt: string;
  /** @nullable */
  paidAt?: string | null;
}