# lib/api-zod/src/generated/types/walletVerifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arete – Solana Pay invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletVerifyInput {
  wallet: string;
  signature: string;
  nonce: string;
}